Move other Data Analysis mock objects to new file
